fix(quality): drop unused imports (S1128), redundant long cast (S1905), add JwtUtils private constructor (S1118)

+++ it-service-backend/src/main/java/com/ticketsystem/it_service_backend/service/SlaPolicyService.java
@@ -100,6 +100,6 @@ public class SlaPolicyService {
     }
 
     private long defaultMs(String priority) {
-        return (long) defaults(priority)[0] * 3_600_000L;
+        return defaults(priority)[0] * 3_600_000L;
     }
 }
